Extract parts of main.go into separate files
The main.go file was getting long, so extracted the cells, color,
and picture logic into separate files.
